third_party/blink/web_tests/fast/dom/elementsFromPoint/elementsFromPoint-svg-expected.txt

PASS nodeListToString(document.elementsFromPoint(125, 125)) is "rect#topLeftRect3, rect#topLeftRect1, svg#svg, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(325, 125)) is "rect#topRightRect2, rect#topRightRect1, svg#topRightSvg, svg#svg, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(225, 225)) is "rect#middleRect2, rect#middleRect1, g#middleG2, g#middleG1, svg#svg, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(125, 325)) is "image#bottomLeftImage2, image#bottomLeftImage1, g#bottomLeftG, svg#svg, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(325, 325)) is "rect#bottomRightRect2, rect#bottomRightRect1, g#bottomRightG2, g#bottomRightG1, svg#svg, DIV#sandbox, BODY, HTML"
PASS nodeListToString(document.elementsFromPoint(350, 250)) is "DIV#foDiv, foreignObject#fo, svg#svg, DIV#sandbox, BODY, HTML"
PASS successfullyParsed is true

TEST COMPLETE
